Update "files" object in package.json (#136)
This commit/PR removes the "*" line of the files object in package.json. That catch-all allowed .gitignore files to be in the released plugin code, which conflicted with the CampusPress deployment workflow and caused the outage/error last week.

This commit/PR adds a more terse ruleset in the files object, which will prevent such issues in the future.